+++ web/src/lib/components/trail/trail_card.svelte
@@ -75,20 +75,25 @@
                 </h5>
             </div>
         </div>
-        <div class="flex mt-1 gap-4 text-sm text-gray-500 whitespace-nowrap">
+        <div class="grid grid-cols-2 mt-2 gap-1 text-sm text-gray-500 whitespace-nowrap">
             <span
                 ><i class="fa fa-left-right mr-2"></i>{formatDistance(
                     trail.distance,
                 )}</span
             >
             <span
-                ><i class="fa fa-up-down mr-2"></i>{formatElevation(
+                ><i class="fa fa-clock mr-2"></i>{formatTimeHHMM(
+                    trail.duration,
+                )}</span
+            >
+            <span
+                ><i class="fa fa-arrow-trend-up mr-2"></i>{formatElevation(
                     trail.elevation_gain,
                 )}</span
             >
             <span
-                ><i class="fa fa-clock mr-2"></i>{formatTimeHHMM(
-                    trail.duration,
+                ><i class="fa fa-arrow-trend-down mr-2"></i>{formatElevation(
+                    trail.elevation_loss,
                 )}</span
             >
         </div>